backend/requirements.txt
fastapi
uvicorn
earthengine-api
python-dotenv
httpx
torch
pandas
scikit-learn
pillow
